Refactor recipe renderer to reuse shared modules.
Remove duplicated Takumi and Satori implementations from recipe-renderer and keep the shared renderer modules aligned with the current rendering behavior.

Made-with: Cursor
