+++ lib/recipes/renderers/satori.ts
@@ -9,12 +9,16 @@ export async function renderWithSatori(
 	width: number,
 	height: number,
 ): Promise<Buffer> {
-	const pngResponse = new ImageResponse(element, {
+	const imageOptions = {
 		width,
 		height,
 		fonts,
+		shapeRendering: 1,
+		textRendering: 0,
+		imageRendering: 1,
 		debug: false,
-	});
+	};
+	const pngResponse = new ImageResponse(element, imageOptions);
 	const pngBuffer = await pngResponse.arrayBuffer();
 	return Buffer.from(pngBuffer);
 }
