+++ lib/recipes/renderers/takumi.ts
@@ -13,7 +13,22 @@ export async function renderWithTakumi(
 ): Promise<Buffer> {
 	const node = await fromJsx(element);
 	const urls = extractResourceUrls(node);
-	const fetchedResources = urls.length > 0 ? await fetchResources(urls) : [];
+	let fetchedResources: Awaited<ReturnType<typeof fetchResources>> = [];
+	if (urls.length > 0) {
+		try {
+			fetchedResources = await fetchResources(urls);
+		} catch (error) {
+			if (
+				process.env.NODE_ENV !== "production" ||
+				process.env.DEBUG === "true"
+			) {
+				console.warn(
+					"Failed to fetch some external resources, rendering without them",
+					error,
+				);
+			}
+		}
+	}
 	const png = await renderer.render(node, {
 		width,
 		height,
